move readers and writers to common directory
