+++ src/block_device_monitor.cpp
@@ -6,8 +6,8 @@
 #include "block_device.hpp"
 #include "block_device_event.hpp"
 #include "netlink.hpp"
-#include "writers/socket_writer.hpp"
-#include "readers/socket_reader.hpp"
+#include "common/writers/socket_writer.hpp"
+#include "common/readers/socket_reader.hpp"
 #include "common/auto_lock.hpp"
 #include "common/auto_mount.hpp"
 #include "common/temp_folder.hpp"
