+++ src/common/writers/console_writer.cpp
