+++ src/common/writers/console_writer.hpp
