+++ src/common/writers/socket_writer.cpp
